requirements.txt
datasets==3.1.0
gdown==5.2.0
jupyter==1.1.1
matplotlib==3.8.0
numpy==1.26.4
open_clip_torch==2.29.0
pillow==11.0.0
scikit-learn==1.5.2
scipy==1.13.1
torch==2.5.1
torchvision==0.20.1
